Update how interests are read from storage
Originally, the checks were done internally within the InterestsList class itself. However, this caused the program to terminate early without creating an empty address book instead. This has been fixed and now an empty address book will be created when there is something wrong with the interests, similar to the other fields.