Implement feature to display first person on the list whenever list changes
So initially when transiting from 'find list' to 'normal list' there was a problem whenever you try to display the same index. For example, if i chose to display index number 5 in the find list, and then hop back over to the normal list and try to display 5 (which may or may not have a different contact at that index now), i would be unnecessarily met with 'index is alr displayed' msg. Therefore i fixed this by resetting the displayed contact to the first one in the list when the list changes.

**Note** Does not affect the functionalities saketh implemented prior which aims to keep the displayed contact constant in the event when contacts are deleted and added