Add the idea of a current predicate
Prevents the list from always resetting to the showAllPersonsPredicate after every command. Particularly for the find command which makes use of predicates as well. Initially, after doing any other command after a find command, it causes the list to reset to its original state. As a result the user needs to key in the find command again